GPS Simulator Refactorized
