+++ mtvs-gps-simulator/dtos/location-message.dto.ts
@@ -1,0 +1,9 @@
+import { LocationDto } from "./location.dto";
+
+export class LocationMessageDto {
+    timestamp: string;
+    location: LocationDto;
+    carId: string;
+    speed: number;
+    status: 'MOVING' | 'STOPPED';
+}
