+++ mtvs-gps-simulator/dtos/location.dto.ts
@@ -1,0 +1,4 @@
+export class LocationDto {
+    latitude: number;
+    longitude: number;
+}
