+++ mtvs-gps-simulator/dtos/traffic-light-color-message.dto.ts
@@ -1,0 +1,5 @@
+export class TrafficLightColorMessageDto {
+  timestamp: string;
+  trafficLightId: string;
+  color: 'RED' | 'YELLOW' | 'GREEN';
+}
